BUG: Fix beam couch angle computation
Since in the current version what is stationary in RAS is the table top instead of the treatment machine body, the couch angle does not need to be applied directly wrt to the room (FixedReference), but wrt to the table top. Hence the couch angle of the beams is inverted.